+++ CHANGES.md
@@ -8,6 +8,7 @@
 * Opening a particle effect file that doesn't have the locally saved particle images asks you to locate them in a new dialog
 * The currently open file is displayed in the window title and indicates if changes have been made
 * Fixed merging particles resulting in the wrong order of particles being added to the particle effect
+* Fixed undoing remove image always adding the image to the end of the list
 * Minor bugfixes and code cleanup
 ### Version 0.0.3
 * Fixed certain keyboard shortcuts being triggered at the same time